+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_client.cpp
@@ -323,10 +323,12 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
   {
     Blobs::UploadBlockBlobFromOptions blobOptions;
     blobOptions.Context = options.Context;
-    blobOptions.ChunkSize = options.ChunkSize;
+    blobOptions.TransferOptions.SingleUploadThreshold
+        = options.TransferOptions.SingleUploadThreshold;
+    blobOptions.TransferOptions.ChunkSize = options.TransferOptions.ChunkSize;
+    blobOptions.TransferOptions.Concurrency = options.TransferOptions.Concurrency;
     blobOptions.HttpHeaders = FromPathHttpHeaders(options.HttpHeaders);
     blobOptions.Metadata = options.Metadata;
-    blobOptions.Concurrency = options.Concurrency;
     return m_blockBlobClient.UploadFrom(fileName, blobOptions);
   }
 
@@ -337,10 +339,12 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
   {
     Blobs::UploadBlockBlobFromOptions blobOptions;
     blobOptions.Context = options.Context;
-    blobOptions.ChunkSize = options.ChunkSize;
+    blobOptions.TransferOptions.SingleUploadThreshold
+        = options.TransferOptions.SingleUploadThreshold;
+    blobOptions.TransferOptions.ChunkSize = options.TransferOptions.ChunkSize;
+    blobOptions.TransferOptions.Concurrency = options.TransferOptions.Concurrency;
     blobOptions.HttpHeaders = FromPathHttpHeaders(options.HttpHeaders);
     blobOptions.Metadata = options.Metadata;
-    blobOptions.Concurrency = options.Concurrency;
     return m_blockBlobClient.UploadFrom(buffer, bufferSize, blobOptions);
   }