+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -21,6 +21,8 @@
 - Removed `PreviousContinuationToken` from `ListBlobContainersSinglePageResult`, `ListBlobsByHierarchySinglePageResult` and `ListBlobsSinglePageResult`.
 - `ListBlobContainersIncludeItem` was renamed to `ListBlobContainersIncludeFlags`.
 - `ListBlobsIncludeItem` was renamed to `ListBlobsIncludeFlags`.
+- `Concurrency`, `ChunkSize` and `InitialChunkSize` were moved into `DownloadBlobToOptions::TansferOptions`.
+- `Concurrency`, `ChunkSize` and `SingleUploadThreshold` were moved into `UploadBlockBlobFromOptions::TransferOptions`.
 - Removed `TagValue` from `FilterBlobItem`, removed `Where` from `FindBlobsByTagsSinglePageResult`.
 - Type for ETag was changed to `Azure::Core::ETag`.
 - Removed `BlobPrefix` struct, use `std::string` instead.